Update handlers.go
Changed HTTP code 301 to 302 in RedirectShortURL() because 301 means the page was permanently moved and the browser remembers this, so it automatically goes to the long URL meaning the backend does not save a new ShortenedURLVisitsHistory. By changing it to 302, the requests will go through the backend every time and let the browser know we found the new URL (long URL).